requirements.txt
numpy
sympy>=1.8
urdfpy
regex
PyYAML